Adding Wireshark-independent decryption.
A large amount of code for the decryption mechanism was
introduced to VMess dissector. The testing cases were
put in a separate repo https://github.com/SIGPET-SEU/CTest.

The newly introduced code passed the compiler's check,
but a lot of warnings concerning the memory and NULL pointer
are shown. I'll check them in the later commits, so DO NOT
try to use these routines DIRECTLY unless you know what you're
doing.