Fix bug.
The original dissector does not be tested on data-encrypted captures, thus the missing
auth will not raises any issue. Once the VMess data is encrypted, missing auth will lead
to Read Violation Error when using NULL for GHashTable lookup.

Therefore, the auth_missing variable as a check is introduced for early return once the
decryption for this conversation is found impossible.

However, the better solution is to try a heuristic dissector since securityNone, or no
encryption might be adopted. I'll implement it in the future.