fix: fixed Trojan tree
Fetching Trojan tree from the parent tree results
in the issue that when there are multiple Trojan
trees, the index besides the first Trojan tree could
not be decided before all TLS tree nodes are constructed.

Therefore, we decide if a tree should be a Trojan
tree after a TLS node is generated. Such a thorough
scan could avoid overlooking other Trojan nodes.